requirement.txt
Django
python-dotenv
google-generativeai
Pillow
Markdown2
opencv-python
numpy
